added ts for teqblaze utils and some fixes (#14680)
