+++ test/spec/libraries/teqblazeUtils/bidderUtils_spec.js
@@ -25,7 +25,8 @@ describe('TeqBlazeBidderUtils', function () {
       bidder: bidder,
       mediaTypes: {
         [BANNER]: {
-          sizes: [[300, 250]]
+          sizes: [[300, 250]],
+          battr: [1, 3]
         }
       },
       params: {
@@ -45,7 +46,8 @@ describe('TeqBlazeBidderUtils', function () {
         [VIDEO]: {
           playerSize: [[300, 300]],
           minduration: 5,
-          maxduration: 60
+          maxduration: 60,
+          battr: [1, 3]
         }
       },
       params: {
@@ -159,8 +161,7 @@ describe('TeqBlazeBidderUtils', function () {
         'tmax',
         'bcat',
         'badv',
-        'bapp',
-        'battr'
+        'bapp'
       );
       expect(data.deviceWidth).to.be.a('number');
       expect(data.deviceHeight).to.be.a('number');
@@ -187,18 +188,17 @@ describe('TeqBlazeBidderUtils', function () {
         expect(placement.type).to.exist.and.to.equal('publisher');
         expect(placement.eids).to.exist.and.to.be.deep.equal(userIdAsEids);
 
-        if (placement.adFormat === BANNER) {
-          expect(placement.sizes).to.be.an('array');
-          expect(placement.gpid).to.be.an('string');
-        }
         switch (placement.adFormat) {
           case BANNER:
             expect(placement.sizes).to.be.an('array');
+            expect(placement.gpid).to.be.an('string');
+            expect(placement.battr).to.exist.and.to.be.deep.equal([1, 3]);
             break;
           case VIDEO:
             expect(placement.playerSize).to.be.an('array');
             expect(placement.minduration).to.be.an('number');
             expect(placement.maxduration).to.be.an('number');
+            expect(placement.battr).to.exist.and.to.be.deep.equal([1, 3]);
             break;
           case NATIVE:
             expect(placement.native).to.be.an('object');
@@ -512,8 +512,16 @@ describe('TeqBlazeBidderUtils', function () {
   });
 
   describe('getUserSyncs', function () {
-    it('Should return array of objects with proper sync config , include GDPR', function () {
+    it('Should return an empty array if no sync enabled', function () {
       const syncData = spec.getUserSyncs({}, {}, {
+        consentString: 'ALL',
+        gdprApplies: true,
+      }, undefined);
+      expect(syncData).to.be.an('array');
+      expect(syncData).to.be.an.deep.equal([]);
+    });
+    it('Should return array of objects with proper sync config , include GDPR', function () {
+      const syncData = spec.getUserSyncs({ pixelEnabled: true }, {}, {
         consentString: 'ALL',
         gdprApplies: true,
       }, undefined);
@@ -525,7 +533,7 @@ describe('TeqBlazeBidderUtils', function () {
       expect(syncData[0].url).to.equal(`https://${DOMAIN}/image?pbjs=1&gdpr=1&gdpr_consent=ALL&coppa=0`)
     });
     it('Should return array of objects with proper sync config , include CCPA', function () {
-      const syncData = spec.getUserSyncs({}, {}, {}, '1---');
+      const syncData = spec.getUserSyncs({ pixelEnabled: true }, {}, {}, '1---');
       expect(syncData).to.be.an('array').which.is.not.empty;
       expect(syncData[0]).to.be.an('object')
       expect(syncData[0].type).to.be.a('string')
@@ -534,7 +542,7 @@ describe('TeqBlazeBidderUtils', function () {
       expect(syncData[0].url).to.equal(`https://${DOMAIN}/image?pbjs=1&ccpa_consent=1---&coppa=0`)
     });
     it('Should return array of objects with proper sync config , include GPP', function () {
-      const syncData = spec.getUserSyncs({}, {}, {}, undefined, {
+      const syncData = spec.getUserSyncs({ pixelEnabled: true }, {}, {}, undefined, {
         gppString: 'abc123',
         applicableSections: [8]
       });
