+++ test/spec/modules/contxtfulBidAdapter_spec.js
@@ -884,8 +884,7 @@ describe('contxtful bid adapter', function () {
         const userSyncs = spec.getUserSyncs({}, [{ body: bidResponse }]);
         expect(userSyncs).to.deep.equal([
           {
-            'url': 'mysyncurl.com/image?pbjs=1&coppa=0&qparam1=qparamv1&qparam2=qparamv2',
-            'type': 'image'
+            'url': 'mysyncurl.com?qparam1=qparamv1&qparam2=qparamv2'
           }
         ]);
       });
